Add color variables, style header
This should've been two commits, really, one for
creating the color variables, the other for
styling the header.

It makes sense to extract the color variables so
managing the palette happens in one place. I created
one set of variables for the values to give them
names and another to link them to the components.

The component list is bound to vary more, but it
still keeps things compact and manageable and allows
the definition of bg/color pairs.

The header section got a few classes. The plan is
to kep each component 50em wide and use margin 0
auto to align them within the background. I hope
that works out for the other components.